# Avoid allocating tintError twice