Passing the type-asserted `tintError` value as an error causes a second allocation.

Since the helper method is only for `tintError` values, change the signature to specify the concrete type to elide the second allocation.