@lmittmann 2024-12-10
I can't verify your claim in the allocs of the benchmark. Can you provide some instructions to verify this?