@database64128 2024-12-10
I wrote this simple benchmark because the full `BenchmarkLogAttrs` seemed a bit much for my laptop, and I'm currently not around my desktop. You should be able to observe the same difference if you add `tint.Err` to `BenchmarkLogAttrs`.

Maybe we should simply replace `slog.Any("error", errTest)` with `tint.Err(errTest)` in `BenchmarkLogAttrs`. What do you think?